Start of second pass, more ARMv7 ASM, and fixes

+++ include/backend/intermediate/symboltable.h
@@ -43,7 +43,7 @@ typedef struct variable_symbol {
         // bool global;
         // bool need_stack;
     u8 uses;
-    u32 stack_location;
+    u16 stack_location;
 } variable_symbol;
 
 /*
